refactor: change file types ty typescript
